data: regenerating based on the latest Swagger
