+++ data/Pandora.Definitions.ResourceManager/ElasticSan/Terraform/ElasticSan-Resource-Schema.cs
@@ -35,13 +35,14 @@ public class ElasticSanResourceSchema
     public CommonSchema.Location Location { get; set; }
 
 
-    [Documentation("The name which should be used for this resource")]
+    [Documentation("Specifies the name of this Elastic San.")]
     [ForceNew]
     [HclName("name")]
     [Required]
     public string Name { get; set; }
 
 
+    [Documentation("Specifies the name of the Resource Group within which this Elastic San should exist.")]
     [ForceNew]
     [HclName("resource_group_name")]
     [Required]
